reworking build scripts... needs testing

+++ bin/build-toolchain.sh
@@ -2,7 +2,9 @@
 
 set -e
 
-TOOLCHAINS_DIR="/opt/toolchains"
+THIS_SCRIPT="$([ -L "$0" ] && readlink -f "$0" || echo "$0")"
+SCRIPT_DIR="$(cd "$(dirname "$THIS_SCRIPT")" ; pwd -P)"
+source "$SCRIPT_DIR/platform.sh"
 
 # ------------------------------------------------------------------------- Help
 
@@ -14,9 +16,12 @@ show_help()
 
    Options:
 
-      --cleanup              Remove temporary files after building
-      --no-cleanup           Do not remove temporary files after building
-      --force                Force reinstall of target
+      --install-dependencies  Instead of building, install build dependencies
+
+      --cleanup               Remove temporary files after building
+      --no-cleanup            Do not remove temporary files after building
+      --force                 Force reinstall of target
+      --toolchain-root        Overwrite default toolchain directory '$TOOLCHAINS_DIR'
 
    Tool:
 
@@ -24,6 +29,9 @@ show_help()
       llvm-x.y.z
 
    Examples:
+
+      # Install build dependencies; requires sudo or root
+      > $(basename $0) --install-dependencies
 
       # Install gcc version 13.1.0 to $TOOLCHAINS_DIR
       > $(basename $0) gcc-13.1.0
@@ -45,20 +53,6 @@ TOOLS_DIR="$TOOLCHAINS_DIR/tools"
 HOST_CC="/usr/bin/gcc"
 HOST_CXX="/usr/bin/g++"
 CMAKE="$TOOLS_DIR/bin/cmake"
-
-if [ -x "/usr/bin/lsb_release" ] ; then    
-    PLATFORM="ubuntu"
-elif [ -f "/etc/fedora-release" ] ; then
-    PLATFORM="fedora"
-elif [ "$(uname -s)" = "Darwin" ] ; then
-    PLATFORM="macos"
-    LIBTOOL=glibtool
-    LIBTOOLIZE=glibtoolize 
-elif [ -f /etc/os-release ] && cat /etc/os-release | grep -qi Oracle  ; then
-    PLATFORM="oracle"
-else
-    echo "Failed to determine platform" 1>&2 && exit 1
-fi
 
 # ------------------------------------------------------------- helper functions
 
@@ -107,9 +101,20 @@ ensure_directory()
 # --------------------------------------------------------- install_dependencies
 
 install_dependences()
-{    
-    if [ "$PLATFORM" = "ubuntu" ] ; then
+{
+    echo "Installing build dependencies"
+    
+    # Don't attempt to use SUDO if it is not installed
+    SUDO="$(which sudo 2>/dev/null || echo "")"
+    if [ "$EUID" -ne "0" ] && [ "$SUDO" = "" ] ; then
+        echo "Either 'sudo' must be installed, or must run as root" 1>&2
+        exit 1
+    fi
+    
+    # TODO: use a fresh docker container to figure our the minimal list
+    if [ "$VENDOR" = "ubuntu" ] || [ "$VENDOR" = "debian" ] ; then
         export DEBIAN_FRONTEND=noninteractive
+        # TODO: test for 'sudo', if ignore this line if it doesn't exist
         sudo apt-get install -y -qq \
              wget subversion automake swig python2.7-dev libedit-dev libncurses5-dev  \
              gcc-multilib python3-dev python3-pip python3-tk python3-lxml python3-six \
@@ -117,11 +122,8 @@ install_dependences()
              libgmp-dev libmpfr-dev libmpc-dev patchelf liblz-dev pax-utils bison flex \
              libxapian-dev
 
-    elif [ "$PLATFORM" = "oracle" ] || [ "$PLATFORM" = "fedora" ] ; then
-        which bison 1>/dev/null 2>/dev/null || sudo yum install -y bison
-        if [ ! -x /usr/bin/xapian-config ] ;then
-           sudo yum install -y xapian-core-devel
-        fi
+    elif [ "$VENDOR" = "oracle" ] ; then
+        sudo yum install -y bison xapian-core-devel
     fi
 }
 
@@ -147,7 +149,8 @@ build_cmake()
 
 build_llvm()
 {
-    local CLANG_V="$1"
+    local INSTALL_PREFIX="$1"
+    local VERSION="$2"
     local LLVM_DIR="llvm"
 
     # First ensure we have a recent cmake
@@ -156,8 +159,7 @@ build_llvm()
     fi
     
     local SRC_D="$TMPD/$LLVM_DIR"
-    local BUILD_D="$TMPD/build-llvm-${CLANG_V}"
-    local INSTALL_PREFIX="${TOOLCHAINS_DIR}/llvm-${CLANG_V}"
+    local BUILD_D="$TMPD/build-llvm-${VERSION}"
     
     rm -rf "$BUILD_D"
     mkdir -p "$SRC_D"
@@ -171,10 +173,10 @@ build_llvm()
     cd llvm-project
     git checkout main
     git pull origin main
-    git checkout "llvmorg-${CLANG_V}"
+    git checkout "llvmorg-${VERSION}"
 
     # local EXTRA_LLVM_DEFINES=""
-    # local MAJOR_VERSION="$(echo "$CLANG_V" | awk -F. '{ print $1 }')"
+    # local MAJOR_VERSION="$(echo "$VERSION" | awk -F. '{ print $1 }')"
     # if (( $MAJOR_VERSION <= 12 )) ; then
     #     EXTRA_LLVM_DEFINES="-D COMPILER_RT_BUILD_GWP_ASAN=Off -D COMPILER_RT_BUILD_LIBFUZZER=Off -D COMPILER_RT_BUILD_ORC=Off -D COMPILER_RT_BUILD_SANITIZERS=OFF"
     #     #-DBacktrace_INCLUDE_DIR=$INSTALL_PREFIX/include -DBacktrace_LIBRARY=$INSTALL_PREFIX/lib/libexecinfo.so
@@ -194,6 +196,7 @@ build_llvm()
          -D LIBCXX_ENABLE_STATIC=YES \
          -D LLVM_BUILD_LLVM_DYLIB=YES \
          -D LLVM_LIBC_ENABLE_LINTING=Off \
+         -D LLVM_USE_LINKER=lld \
          -D COMPILER_RT_ENABLE_IOS:BOOL=Off \
          $EXTRA_LLVM_DEFINES \
          -D CMAKE_INSTALL_PREFIX:PATH="$INSTALL_PREFIX" \
@@ -208,11 +211,10 @@ build_llvm()
 
 build_gcc()
 {
-    local TAG="$1"
-    local SUFFIX="$1"
-    if [ "$2" != "" ] ; then SUFFIX="$2" ; fi
+    local INSTALL_PREFIX="$1"
+    local VERSION="$2"
     
-    local MAJOR_VERSION="$(echo "$SUFFIX" | sed 's,\..*$,,')"
+    local MAJOR_VERSION="$(echo "$VERSION" | sed 's,\..*$,,')"
     local SRCD="$TMPD/$SUFFIX"
     
     mkdir -p "$SRCD"
@@ -223,7 +225,7 @@ build_gcc()
     
     cd gcc
     git fetch
-    git checkout releases/gcc-${TAG}
+    git checkout releases/gcc-${VERSION}
 
     if (( $MAJOR_VERSION <= 9 )) ; then
         # some network configurations have trouble with 'ftp://' urls
@@ -235,17 +237,16 @@ build_gcc()
     mkdir -p "$SRCD/build"
     cd "$SRCD/build"
 
-    local PREFIX="${TOOLCHAINS_DIR}/gcc-${SUFFIX}"
-
     export CC=$HOST_CC
     export CXX=$HOST_CXX
     nice ../gcc/configure \
-         --prefix=${PREFIX} \
+         --prefix=${INSTALL_PREFIX} \
          --enable-languages=c,c++,objc,obj-c++ \
          --disable-multilib \
          --program-suffix=-${MAJOR_VERSION} \
          --enable-checking=release \
-         --with-gcc-major-version-only
+         --with-gcc-major-version-only \
+         --disable-nls
     
     nice make -j$(nproc) 2>$SRCD/build/stderr.text | tee $SRCD/build/stdout.text
     nice make install | tee -a $SRCD/build/stdout.text
@@ -260,6 +261,9 @@ done
 
 CLEANUP="True"
 FORCE_INSTALL="False"
+INSTALL_DEPS="False"
+TMPD=""
+COMPILER=""
 
 while (( $# > 0 )) ; do
     ARG="$1"
@@ -267,30 +271,37 @@ while (( $# > 0 )) ; do
     [ "$ARG" = "--cleanup" ]       && export CLEANUP="True" && continue
     [ "$ARG" = "--no-cleanup" ]    && export CLEANUP="False" && continue
     [ "$ARG" = "--force" ] || [ "$ARG" = "-f" ] && export FORCE_INSTALL="True" && continue
-
+    [ "$ARG" = "--install-dependencies" ] && INSTALL_DEPS = "True" && continue
+    [ "$ARG" = "--toolchain-root"  && export TOOLCHAINS_DIR="$1" && shift && continue
+    
     if [ "${ARG:0:3}" = "gcc" ] ; then
-        COMMAND="build_gcc ${ARG:4}"
-        CC_MAJOR_VERSION="$(echo ${ARG:4} | awk -F. '{ print $1 }')"
-        EXEC="$TOOLCHAINS_DIR/$ARG/bin/gcc-$CC_MAJOR_VERSION"        
+        COMPILER="gcc"
+        VERSION="${ARG:4}"
+        CC_MAJOR_VERSION="$(echo ${VERSION} | awk -F. '{ print $1 }')"
+        EXEC="gcc-$CC_MAJOR_VERSION"        
         continue
     fi    
 
     if [ "${ARG:0:4}" = "llvm" ] || [ "${ARG:0:5}" = "clang" ] ; then
         [ "${ARG:0:4}" = "llvm" ]  && VERSION="${ARG:5}" || true
         [ "${ARG:0:5}" = "clang" ] && VERSION="${ARG:6}" || true
-        COMMAND="build_llvm $VERSION"
-        EXEC="$TOOLCHAINS_DIR/llvm-$VERSION/bin/clang"
+        COMPILER="llvm"
+        EXEC="clang"        
         continue
     fi  
 
     echo "unexpected argument: '$ARG'" 1>&2 && exit 1
 done
+      
+if [ "$INSTALL_DEPS" = "True" ] ; then
+    install_dependences
+fi
 
 if [ "$COMMAND" = "" ] ; then
     echo "Must specify a build command!" 1>&2 && exit 1
 fi
 
-SCRIPT_NAME="$(basename "$0")"
+SCRIPT_NAME="$(basename "$THIS_SCRIPT")"
 if [ "$CLEANUP" = "True" ] ; then
     TMPD="$(mktemp -d /tmp/$(basename "$SCRIPT_NAME" .sh).XXXXXX)"
 else
@@ -298,21 +309,20 @@ else
     mkdir -p "$TMPD"
     echo "Setting TMPD=$TMPD"
 fi
-
 trap cleanup EXIT
+
 cleanup()
 {
-    if [ "$CLEANUP" = "True" ] ; then
+    if [ "$CLEANUP" = "True" ] && [ "$TMPD" != "" ] ; then
         rm -rf "$TMPD"
     fi
 }
 
 # ----------------------------------------------------------------------- action
-
-if [ "$FORCE_INSTALL" = "True" ] || [ ! -x "$EXEC" ] ; then
+INSTALL_PREFIX="${TOOLCHAINS_DIR}/${COMPILER}-${VERSION}--${MACHINE}--${VENDOR_TAG}--${PLATFORM}"
+if [ "$FORCE_INSTALL" = "True" ] || [ ! -x "$INSTALL_PREFIX/bin/$EXEC" ] ; then
     ensure_directory "$TOOLCHAINS_DIR"
-    install_dependences
-    $COMMAND
+    build_$COMPILER "$INSTALL_PREFIX" $VERSION
 else
     echo "Skipping installation, executable found: '$EXEC'"
 fi
